+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unpublished
 
+* Fix build of zng-view-* without `"ipc"` feature.
+* Prebuilt view-process now uses the same tracing context as the app-process.
+    - Note that the tracing context must be set before `run_same_process`.
+* Fix "GLXBadWindow" fatal error in Ubuntu.
 
 # 0.7.0